Ungraceful fallback for scrollIntoView
Some browsers do not support the parameters for scrollIntoView, accepting a Boolean only.  There is no guaranteed method to check if parameters are supported (a particular property appears closely related but using it for detection is an excercise in presumptions), thus using a try/catch.  
The fallback will use the Boolean value 'false', which is equivalent to scrolling to the end of the document vertically, while still using "nearest" horizontally.