Fix themes bounds check being off-by-one (#323)
The offending line was in checking the index to be strictly greater than
the length rather than greater than or equal to. This means the app
would not load if you were using a theme of an index exactly one higher
than the maximum.

The rest of the changes just clean the code up to handle strings and
other junk.