Implement workaround for Chrome blurring background with filter() (#345)
This involves layering another canvas element on top of the board. For
unknown reasons, this has resulted in Chrome correctly using pixelated
rendering for the board in tests. While this is ineffective for truely
unrendered elements, we can trick chrome into thinking the fix element
is being used by giving it a gradient background which starts at a
non-transparent value and immediately becomes transparent at 0%.