Fix ban element showing multiple close buttons. (#351)
Commit 2c068d7bf7c8b9f500cefbe7bf2419cac9f66707 changed modal defaults
and opts to always show a close button regardless of passed options. As
a side effect, the modals which do opt to show the stock close button
now display two anchors.
I've removed the ban modal's overriding boolean, but work is required in
the future to actually honor modal options. This commit is essentially a
bandaid in preparation for reset.