Merge heatmap and virginmap code (#346)
* Consolidate heatmap and virginmap into single "overlays" namespace

Each of these is now implemented as an `overlay` object. Overlays have a
unique name and take functions in their constructor allowing the
constructing code to specify:
- How the initial data should be created when the overlay is first
  shown.
- What should happen after the initial data has been loaded.

* Add reload function to overlays

This will run lazyInit for the overlay so long as it is not currently
running. The callback for when lazyInit completes is now passed a
boolean to let the callback know if this is the first load or not so as
not to re-register listeners and such.

* Remove unused virginmap element

* Remove redundant attribute assignment

* Remove code setting old ls key for notification sound

* Fix heatmap background color being green instead of black