Role Overhaul (#389)
* Role Overhaul

This commit brings way too many changes for me to remember them all, but nonetheless:

* Added support for users to have multiple roles (role -> roles VARCHAR(512), role IDs are delimited by comma)

* Added roles-reference.conf as a reference role configuration

* Polished console commands related to banning and role assignment

* Removed banned and shadowbanned roles in favor of ban_expiry and (new) is_shadow_banned column checks respectively

* Delegated badges to role configuration

* Add permission gate on most endpoints and packets

* Remove role types

After further inspection, the role type property can be removed. The guest role will now use the `guest` boolean property, which goes hand-in-hand with the `user` role which uses the `default` boolean property.

* Remove redundant Permission and Role managers

Removes Permission.java, PermissionManager.java, and RoleManager.java in favor of static methods on Role.java.

* Remove redundant user ban check

* Simplify setOverrideCooldown

* Fix ban expiry misinterpretation and shadow-bans

* Remove unused board.place.localonly permission

* Removed unused permissions HashMap

* Remove unused permissions HashMap

* Add README step for copying role configuration

* Add roles table and fix client user permissions

This commit adds a `roles` table in place of the `roles` column on `users`, and fixes the `user.hasPermission` method (pxls.js) not including direct permissions.

* Remove chat.staffRainbow config property

Obsoleted by permissions.

* Add and update README console commands

* Tweak roles.md description

* Add role list to profile

* Give user default roles on sign-up

This also changes all instances of Database.setUserRoles to pass a user ID instead of User.

* Fix JavaDoc inconsistency

* Reload roles on reload command

* Allow role name and ID input in console commands

* Add roles.conf to .gitignore

roles.conf does not exist, but to stay consistent with the other runtime entries it's added to the .gitignore.

* Separate roles deletion and insertion handles

This is an attempt to solve a report of a duplicate key error.

* Nullify user ban expiry past expiry time

* Reload user manager on reload command

* Remove default roles being assigned in database

* Update reference roles and add guest inheritance

* Developers are no longer considered staff by default. They retain the ability to use a rainbow chat name.

* Added a Trial Moderator role.

* Moved most staff-based permissions to their relevant role (e.g. user.shadowban, all notification permissions to Administrator) rather than all on Staff. Basic staff-only permissions (such as mod checks, rate-limit bypassing, and receiving staff broadcasts) remain on Staff.

* By default, all visitors will inherit guest roles regardless of whether they are signed in.

* List inherited roles in userinfo packet

* Polish roles command

* Send data to shadowbanned users

Sends role updates and ban expiry updates to shadow-banned users. Leftover from SHADOWBANNED logic.

* Update roles.md

* Removed `type`, added `guest`, and fixed `default` type.

* Comment out DBPixelPlacement.roles

It was unused and causing an unnecessary database query.

* Remove development comments

* Remove commented lookup rolegate logic

* Rename ServerUserInfo.getRole to getRoles

* Simplify User.isBanned logic

It actually was just this simple! No longer makes a call to update the database if the user's ban expiry time has passed.

* Remove unused note

* Fix previously banned users counted as banned

* Fix permission check not including guest and default role permissions when the user has more roles assigned

* Use roles-reference.conf as a default when no roles.conf exists, instead of merging it as a fallback.
This way, roles defined in roles-reference.conf can be removed.

* Filter out null roles from Database.getUserRoles()
These can appear due to a role being added to an user but later removed from roles.conf

* Hacky fix to flatmap permissions one final time

Figure out why permissions aren't being flatmapped later.

Co-authored-by: Martín (Netux) Rodríguez <send.email+git@netux.site>