Second round of changes to improve the obfuscation of usernames and QoL improvements for staff during snip mode (#466)
* Hide pixel counts on profile page when snip mode is active

* Send null pixel counts and discord names when snip mode is enabled.

* Send non-redacted chat messages to staff members when snip mode is enabled.
The messages are redacted on the client side to allow staff members to enjoy snip mode; but also keeping vital information like usernames for them to see when clicking on "-snip-" usernames.

This also shifts the way recognizable information in snip mode is redacted out of the objects containing the information itself.
From now on these objects contain an `object.asSnipRedacted()` method which recreates the object with some values redacted. The logic outside of that object's constructor must check `App.getSnipMode()` and call `object.asSnipRedacted()` when necessary.

* Disallow purging during snip mode
Hides "Purge User" button and returns early from the /admin/chatPurge endpoint during snip mode.
This is more of a temporary patch to prevent staff members from accientally purging users, as it'd remove all chat messages on the client side.

* Show warning on chat ban dialog stating that purging is disabled during snip mode

* Don't link to profiles when snip mode is enabled where -snip- is expected as the username.
Concretely: this removes the links on lookups and the Profile button on the chat message popup.