Shadow-Banned Chat Messages v2 (#473)
* Fix broken main tag

* Add support for shadowbanned messages

This commit adds a not null boolean `shadow_banned` column to the
`chat_messages` table to determine whether or not the user who sent the
message was shadow-banned at the time.

If a message is retrieved that has the `shadow_banned` column set to true and
the retrieving user also has the new `chat.history.shadowbanned` permission,
they will be able to see shadow-banned message(s) in slightly smaller
and red text. If the retrieving user does not have that permission, they
will not see any shadow-banned messages.

One side-effect that was not addressed is that if the shadow-banned user
reloads the page or opens the chat on a new browser page, their messages
will not appear. This is like how pixel shadow-bans are currently handled.
It would be best to address both issues in a separate commit.

* Add chat.history.shadowbanned documentation

* Hide shadow-banned messages from non-staff

* Brighten shadowbanned chat message color

* Fix packet order and add comments for clarity

* Do not send separate staff packets if null

* Hide isAuthorShadowBanned in chat history

This commit fixes the isAuthorShadowBanned property from showing up if
the message comes from someone who is not shadow-banned.

This commit also adds a missing JavaDoc param and skips ChatMessage
instantiation if the author is shadow-banned and the includeShadowBanned
method argument is false.

* Remove unused imports in App.java

* Broadcast chat message based on null check

* Remove createChatMessage shadowBanned parameter

* Fix case for getLastXMessagesForSocket

* Move shadow-ban message retrieval logic

This commit moves shadow-ban message retrieval logic to
handleChatHistory instead of in getLastXMessagesForSocket.

* Filter shadow-banned message history

This commit filters chat message history to remove chat messages from
shadow-banned users if the requesting user does not have permission to
read shadow-banned messages in history.

* Improve handling of chat history, specially regarding shadow bans
- Merges logic from `Database#getLastXMessageForSocket()` into `PacketHandler#handleChatHistory()`
- Makes use of the shadow_banned (author_was_shadow_banned) property of chat messages
- Fixes logic error that caused only the shadow banned user to receive their own messages

Overall, this massages Flying's code so that messages from shadow banned users work this way:
- Shadowbanned users receive their own messages, without the shadow ban flag
- Users with both the user.receivestaffbroadcasts and chat.history.shadowbanned permissions receive the message flagged with the shadow ban flag
- Users that don't match any of the above don't receive the message at all
- If the user that sent the message is unbanned later, the messages they sent while shadow banned preserve the shadow ban flag behaviour mentioned above.

Co-authored-by: Andrew Black <andrewblack2656@gmail.com>