Break pxls.js into modules (#508)
* Browserify js source files in gulp

This excludes `resources/public/include/` since they need not be
compiled separately. There are no files there yet, but this should save
some compile time when there are.

Notably, this does not mean that those files are excluded from being
output. The way the current build system seems to work is:
- Maven copies `resources/` to `target/resources/`
- Gulp generates various artifacts and replaces those in
  `target/resources/`

This means that even files unprocessed by gulp end up in the output —
such as the contents of `resouces/public/include/`. Something to
investiagte later, perhaps.

* Move helper functions into separate module

* Move typeahead to separate module

* Move settings to separate module

* Move serviceWorkerHelper to separate module

The naming of this is all a bit questionable — perhaps some renaming is
in order later.

* Move panels to separate module

* Move modal to separate module

* Move everything into separate modules

You may notice that this is a drastic step-up from the previous commits.
This is because the previous new modules had a nice dependency tree.
The rest of the modules (the ones added in this commit) contain a lot of
circular dependencies.

Previously, circular dependencies worked thanks to JavaScript's scoping.
Since we can no longer take advantage of that, some dependencies are
delayed until the module's `init` function is called.

Ideally, there would be no such circular dependencies but that sort of
re-architechting is non-trivial and out of scope for this.