Improve chat ratelimit configurabilty (#522)
* Add a permission for ignoring chat ratelimits

* Make chat cooldown message configurable

Also uses moment for relative time display. This is debatably a
downgrade since it's much less accurate, but I couldn't see any better
ways to do this simply. It's probably better than what the code was
doing before. I'm not sure what `(new Date()) >> 0` evaluates to
exactly, but it doesn't seem to be `valueOf` or `Date.now()`. It does
seem like it's probably in milliseconds, so there's that at least.