Implement localization for script strings (#518)
* Pass all user-facing strings to the function `__`

This function just returns the input with no changes. This function is a
marker for translatable strings and may later fetch string translations.

The common name for this function is usually just `_` (a single
underscore) but since a common practice is to use this for the name of
an unused variable, I felt using two underscores was a bit safer.

Another possible name for this function is `gettext` but this is rather
verbose for what is needed here, given the frequency of this functions
appearence.

* Extract script strings

This could probably have been done with a regex approach as with pebble
templates, but — since I have a JS parser — it is done using esprima.
This is probably more robust while also more likely to be incorrect. It
seems to work fine but I didn't check it all that hard.

Another option to this entirely is to use an established library for
gettext, such as ttag. That would offer a number of benefits including a
plan to actually load translations back into scripts that this currently
doesn't offer. We will see, I suppose…

* Mark strings as needing translations

These were previously missed.

* Refactor and separate JS translation parsing

This is in preparation to re-used the code in gnerating translated
scripts.

* Compile translated script files on build

This all seems to work suprisingly well. The most worrying thing
currently is that this code makes a lot of assumptions about the
language codes used. Notably, it basically pretends that subtypes don't
exist and expects to be able to check for just the first part.

That's not the only thing worrying about all this, but what about pxls
codebase isn't?